+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 Notable changes to this project.
 
 ## dev
+- model_upload now allows specifying which data_version and which image to use
 - added `round_model_performances_v2`, which allows fetching newer metrics
 - added `pipeline_status` to get the status of numerai's scoring pipeline
 - added `intra_round_scores` to fetch intra-round metrics of your models